Fix spam log messages about sticky faults on Spark
The code looks like it's intended to only cause a log message if the state of
sticky faults changes in some way.

getStickyFaults() retuns a new java object each time, so using == for
comparison on the object will always return true.

This change looks at the rawBits of the fault object instead.

old behavior = log message once a second even if there are no sticky faults
new behavior = log message only when the bits actually change

It might arguably be better to log once a second but only when there are faults
so that the presence of faults is easier to see in a busy log file.